eggs_hotspots_week

this function predicts the number of eggs using a geostatistical analysis with INLA & calculates the hotspots of the eggs

eggs_hotspots_week(
  cve_mpo,
  cve_edo,
  locality,
  hist_dataset,
  year,
  integration_strategy,
  aproximation_method,
  fam_distribution,
  path_vect,
  path_coord,
  palette.viridis,
  plot,
  cell.size,
  alpha.value,
  kvalue
)

Arguments

cve_mpo

is the text id of the municipality.

cve_edo

is the text id of the state.

locality

locality is the locality target.

hist_dataset

is a logical value for define the dataset, if TRUE is the ovitraps historical dataset and we neen define the year to analyze.

year

is the year to analyze.

integration_strategy

is the integration strategy. The options are "grid", "eb" & "ccd"

aproximation_method

aproximation is the aproximation of the joint posterior of the marginals and hyperparameter. The options are "gaussian", "simplified.laplace" & "laplace"

fam_distribution

is the name of the family of the distribution for modelling count data. The option can be poisson, zeroinflatedpoisson0, zeroinflatedpoisson1, nbinomial, zeroinflatednbinomial0 and zeroinflatednbinomial1.

path_vect

is the directory of the ovitrampas dataset.

path_coord

is the directory of the coordinates dataset.

palette.viridis

is the palette of viridis. The options are plasma, viridis, inferno & magma.

plot

is a logical value for the plot the mesh.

cell.size

is the sample number per location for predictions.

alpha.value

alpha The significance level, also denoted as alpha or α, is the probability of rejecting the null hypothesis when it is true.

kvalue

is the parameter for define the triagulization of delauney in the inner and the outer area in the argument max.edge in the INLA:inla.mesh.2d.

Value

a list of two object (a df of betas & a list of spde of each week)

Details

inla, eggs_hotspots

Author

Felipe Antonio Dzul Manzanilla felipe.dzul.m@gmail.com

Examples

1+1
#> [1] 2