eggs_map

The function generates an egg map of Aedes Aegypti.

eggs_map(
  path_lect,
  path_coord,
  path_shp,
  leg_title,
  loc,
  size,
  alpha,
  palette_vir,
  risk,
  weeks
)

Arguments

path_lect

is the directory of ovitraps dataset.

path_coord

is the directory of coordinates dataset.

path_shp

is the directory of shepefile dataset.

leg_title

is title of legend.

loc

is the target locality.

size

is the size of point.

alpha

is the transparency of point.

palette_vir

is the name of pallete of viridis.

risk

is logical value, if is TRUE return of percentil map, else only map of eggs.

weeks

is the week number.

Value

a ggplot.

Author

Felipe Antonio Dzul Manzanilla felipe.dzul.m@gmail.com

Examples

1 +1
#> [1] 2