spde_pred_map

This function predicts the number of eggs in areas where it was not collected using geostatistical analysis with INLA

spde_pred_map(
  path_lect,
  locality,
  path_coord,
  cve_ent,
  leg_title,
  alpha,
  plot = NULL,
  aproximation,
  integration,
  longitude,
  latitude,
  k,
  week,
  var,
  cell_size,
  palette_vir
)

Arguments

path_lect

is the directory of the ovitrampas readings file.

locality

is the locality target.

path_coord

is the directory of the ovitrampas coordinates file.

cve_ent

is the text id of the state.

leg_title

is title of legend.

alpha

The significance level, also denoted as alpha or α, is the probability of rejecting the null hypothesis when it is true.

plot

is a logical argument. if TRUE plot the mesh else no plot the mesh.

aproximation

is the aproximation of the joint posterior of the marginals and hyperparameter. The options are "gaussian", "simplified.laplace" & "laplace".

integration

is the integration strategy. The options are "grid", "eb" & "ccd".

longitude

is the name of the column of the longitude in the ovitrampas dataset.

latitude

is the name of the column of the longitude in the ovitrampas dataset.

k

is the parameter for define the triagulization of delauney.

week

is the week target. proactive is the current week.

var

is the name of variable target.

cell_size

is the parameter for define the grid of locality for prediction.

palette_vir

is the palette of viridis. The option can be magma, plasma, inferno, and viridis.

Value

a list with the gmap with the prediction of the number of eggs, the prediction of number of eggs, the mesh.

Details

inla.

See also

Author

Felipe Antonio Dzul Manzanilla felipe.dzul.m@gmail.com

Examples

1+ 1
#> [1] 2